collection_module/requirements.txt
python-dotenv>=0.19.0
requests>=2.26.0
pandas>=1.3.0
openpyxl>=3.0.7
python-magic>=0.4.24
pathlib
zipfile36
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0